Move grpc server impl to separate module to reduce jar size
