+++ grpc/src/main/resources/META-INF/services/org.wso2.msf4j.grpc.ServerBuilder
@@ -1,0 +1,1 @@
+org.wso2.msf4j.grpc.ServerDefinitionBuilder
