+++ pom.xml
@@ -104,6 +104,7 @@
         <module>jaxrs-delegates</module>
         <module>core</module>
         <module>swagger</module>
+        <module>grpc</module>
         <module>spring</module>
         <module>deployer</module>
         <module>poms/msf4j-service</module>
