+++ samples/grpc-stockquote/stockquote-service/README.md
@@ -12,6 +12,47 @@ the microservice.
 Your POM can inherit from [msf4j-service](../../../poms/msf4j-service). 
 See details [here](../../../poms/msf4j-service).
 
+Add following dependencies and plugin definition to the pom file
+
+````xml
+    <build>
+        <extensions>
+            <extension>
+                <groupId>kr.motd.maven</groupId>
+                <artifactId>os-maven-plugin</artifactId>
+                <version>${os.maven.plugin.version}</version>
+            </extension>
+        </extensions>
+        <plugins>
+            <plugin>
+                <groupId>org.xolstice.maven.plugins</groupId>
+                <artifactId>protobuf-maven-plugin</artifactId>
+                <configuration>
+                    <protocArtifact>com.google.protobuf:protoc:3.4.0:exe:${os.detected.classifier}</protocArtifact>
+                    <pluginId>grpc-java</pluginId>
+                    <pluginArtifact>io.grpc:protoc-gen-grpc-java:1.7.0:exe:${os.detected.classifier}
+                    </pluginArtifact>
+                </configuration>
+                <executions>
+                    <execution>
+                        <goals>
+                            <goal>compile</goal>
+                            <goal>compile-custom</goal>
+                        </goals>
+                    </execution>
+                </executions>
+            </plugin>
+        </plugins>
+    </build>
+
+    <dependencies>
+        <dependency>
+            <groupId>org.wso2.msf4j</groupId>
+            <artifactId>msf4j-grpc</artifactId>
+        </dependency>
+    </dependencies>
+````
+
 ## Writing service implementation
 
 Writing gRPC service is similar to the normal microservice. Following are the additional things we need to consider,
